requirements.txt
# Development dependencies
flake8>=6.0.0
isort>=5.0.0
pytest-cov>=4.0.0
black>=23.11.0

# Core dependencies
fastmcp>=0.2.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pytest>=7.4.3
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.8.0
uvicorn>=0.24.0
fastapi>=0.104.1
pydantic>=2.5.1
aiohttp>=3.9.1
anyio>=4.1.0 